created profile photo generator class
