+++ code/noRAM/app/src/main/res/layout/fragment_organizer_event_list.xml
@@ -11,6 +11,7 @@
         android:id="@+id/allEventsButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_margin="10dp"
         android:background="@drawable/button_background"
         android:minWidth="0dp"
         android:minHeight="0dp"
@@ -18,8 +19,7 @@
         android:paddingTop="10dp"
         android:paddingEnd="30dp"
         android:paddingBottom="10dp"
-        android:layout_margin="10dp"
-        android:text="@string/all_events_button"
+        android:text="@string/my_events"
         android:textAllCaps="false"
         android:textColor="@color/button_text_color"
         app:layout_constraintBottom_toTopOf="@+id/allEventsList"
